move tts autoplay settings to instructionRedux
