+++ apps/src/templates/instructions/TopInstructions.story.jsx
@@ -4,14 +4,12 @@ import {createStore, combineReducers} from 'redux';
 import * as commonReducers from '@cdo/apps/redux/commonReducers';
 import {
   setHasAuthoredHints,
-  setInstructionsConstants
+  setInstructionsConstants,
+  setTtsAutoplayEnabled
 } from '@cdo/apps/redux/instructions';
 import {enqueueHints, showNextHint} from '@cdo/apps/redux/authoredHints';
 import isRtl, {setRtlFromDOM} from '@cdo/apps/code-studio/isRtlRedux';
 import {setPageConstants} from '@cdo/apps/redux/pageConstants';
-import sectionData, {
-  setTtsAutoplayEnabled
-} from '@cdo/apps/redux/sectionDataRedux';
 import TopInstructions from './TopInstructions';
 
 /**
@@ -24,9 +22,7 @@ import TopInstructions from './TopInstructions';
  * @param {boolean} options.tts
  */
 const createCommonStore = function(options = {}) {
-  const store = createStore(
-    combineReducers({...commonReducers, isRtl, sectionData})
-  );
+  const store = createStore(combineReducers({...commonReducers, isRtl}));
   const pageConstants = {};
   const instructionsConstants = {};
